The ram usage is not anymore a "pro" Firefox argument
Since Quantum the ram usage is about the same, Firefox sandbox is not as evolved as Chome's which means it might take even more with more features. 

The last official Benchmark with Firefox Vs. Chrome was with FF 50 (or around 50 according to the .moz page) I think they are not updating it because they know exactly that RAM is no argument anymore. Of course, it's only a valid statement if you compare Default vs. default Browser, in case you use lots of "fancy" extensions and settings I guess Chrome is better in the overall consumption because it simply does not has as many third-party libraries (which need to be loaded into ram etc.).