Fix Chrome/Chromium/Edge & Brave's scrollbar + Updates for v76/77/78
The flags are not in any connection with any privacy or security "tweak". This basically is a workaround for several (already reported) bugs in v76/77 & 78+Dev in Edge, Brave, Chrome/Chromium (and possible other forks).

Added a category which shows only "security" related tweaks, for those who only want to change those flags.