Added "use-angle" flag
Oh hey Chrome, after 4 years, the flag is finally introduced AND working. 

It's a Windows only flag which reduces system resources on nVidia systems. Via [OpenGL or ANGLE](https://github.com/mrdoob/three.js/wiki/How-to-use-OpenGL-or-ANGLE-rendering-on-Windows). The flag already got introduced a year ago but was broken, now it finally works.