Keyboard Privacy extension added
Keyboard scrambler vs an extension

The question...
Keyboard scrambler works OS wide also for applications like KeePass as well as Bitwarden, UWP apps and basically all other applications with input fields.

This extension works only in the Browser. But it's more than enough to use this alone, because a keyboard scrambler program often slows down your input, which means it will decrease the overall input delay. Which is bad for e.g. online games when input delay can make the difference. It often then ends up disabling or uninstalling the program, not to mention that the driver itself can cause some problems. The known professional scrambler programs are also never free nor FOSS.

The answer....
I say the extension is enough if you operate in the Browser.